refactor: split device config into srp modules
